Move variable panel hooks to nested folder
