+++ src/components/VariablePanel/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useContentPosition';
+export * from './useTable';
