+++ src/components/VariablePanel/hooks/useContentPosition.test.tsx
