+++ src/components/VariablePanel/hooks/useContentPosition.ts
