+++ src/components/VariablePanel/hooks/useFavorites.test.ts
