+++ src/components/VariablePanel/hooks/useFavorites.ts
