+++ src/components/VariablePanel/hooks/utils.test.ts
