+++ src/components/VariablePanel/hooks/utils.ts
@@ -1,7 +1,7 @@
 import { FilterFn } from '@tanstack/react-table';
 import { PanelData, DataFrame, Field } from '@grafana/data';
 import { locationService } from '@grafana/runtime';
-import { TableItem, Level, RuntimeVariable } from '../../types';
+import { TableItem, Level, RuntimeVariable } from '../../../types';
 
 /**
  * Convert Data Frame to objects array
